more refactoring. stringhelper/error should equal less lines
stringhelper contains common string functions that normally take
up multiple lines of boilerplate

error enull displays which argument is null
new return_ifnull() macro should cut down on if (!arg) log_... bplate

+++ stringarray.h
@@ -17,7 +17,7 @@ struct string_array{
 };
 
 int sa_add(struct string_array* array, const char* str);
-int sa_remove(struct string_array* array, int index);
+int sa_remove(struct string_array* array, size_t index);
 int is_directory(const char* path);
 size_t sa_sanitize_directories(struct string_array* array);
 void sa_sort(struct string_array* array);
